Split the fighter_details table into fighter_tickets and fighter_favor tables
