+++ src/database/queries/ultimate/save-selected-fighters-queries.ts
@@ -1,28 +1,7 @@
-export const getSaveSelectedFightersQuery = (fighterCount: number) => {
-  switch (fighterCount) {
-    case 1:
-      return `
-        UPDATE fighters
-        SET selected = true
-        WHERE fighter_id = ANY($1);
-      `;
-    case 3:
-      return `
-        UPDATE fighters
-        SET selected_ss3 = true
-        WHERE fighter_id = ANY($1);
-      `;
-    case 5:
-      return `
-        UPDATE fighters
-        SET selected_ss5 = true
-        WHERE fighter_id = ANY($1);
-      `;
-    default:
-      return `
-        UPDATE fighters
-        SET selected_x = true
-        WHERE fighter_id = ANY($1);
-      `
-  }
-}
+export const saveSelectedFightersQuery = `
+  UPDATE fighter_tickets
+  SET selected = true
+  WHERE player_id = $1
+    AND fighter_id = ANY($2)
+    AND pool_size = $3;
+`;
