+++ src/database/repos/ultimate-repo.ts
@@ -1,23 +1,23 @@
 import { IDatabase, IMain, ParameterizedQuery } from "pg-promise";
-import { getFightersQuery } from "../queries/ssbu/get-fighters-query";
-import { getSaveSelectedFightersQuery } from "../queries/ssbu/save-selected-fighters-queries";
+import { getFightersQuery } from "../queries/ultimate/get-fighters-query";
+import { saveSelectedFightersQuery } from "../queries/ultimate/save-selected-fighters-queries";
 
 export class UltimateRepository {
   constructor(private db: IDatabase<any>, private pgp: IMain) {}
 
-  async getFighters(fighterCount: number): Promise<any> {
+  async getFighters(playerId: number, poolSize: number): Promise<any> {
     const getFightersPQ = new ParameterizedQuery({
       text: getFightersQuery,
-      values: [fighterCount]
+      values: [playerId, poolSize]
     });
 
     return await this.db.many(getFightersPQ);
   }
 
-  async saveSelectedFighters(fighterIds: number[], fighterCount: number): Promise<any> {
+  async saveSelectedFighters(playerId: number, fighterIds: number[]): Promise<any> {
     const saveSelectedFightersPQ = new ParameterizedQuery({
-      text: getSaveSelectedFightersQuery(fighterCount),
-      values: [fighterIds]
+      text: saveSelectedFightersQuery,
+      values: [playerId, fighterIds, fighterIds.length]
     });
 
     return await this.db.none(saveSelectedFightersPQ);
